requirements.txt
yt-dlp==2025.9.23
PyYAML>=6.0
